+++ NEWS.md
@@ -1,3 +1,13 @@
+# inlmisc 0.4.3.9000
+
+- For leaflet-search plugin, change version from `2.8.0` to `2.9.6` and tidy wrapper functions.
+
+- In `CreateWebMap` function, change URL for TNM basemap WMS services from
+  `"https://basemap.nationalmap.gov/arcgis/services/USGSTopo/MapServer/WmsServer"` to
+  `"https://basemap.nationalmap.gov/ArcGIS/rest/services/USGSTopo/MapServer/tile/{z}/{y}/{x}"`.
+
+- Change required R version from `>= 3.2.0` to `>= 3.4.0`.
+
 # inlmisc 0.4.3
 
 - Add `AddNorthArrow` function, was previously defined as an internal function.